pages.js
The JS for the view. Thanks to regex, first, we will see if the user has typed correctly the expression (for instance, 1-3, 5). Later, we will check if the user want to download a page bigger the document or the page zero. And finally, we will check that the expression has not duplicate elements.